Explicit Mockingbird Generation Deps
Summary:
This is the fbsource copy of D57723694....

When generating mocks if the superclass of the class we're making a mock for lives in another target mockingbird searches through the entire recursive tree of exported deps to find it. This is consuming a very large amount of memory.
This instead requires that when a mock definition is created the author must specify these deps explicitly and by default we don't give mockingbird a tree to look through.

Differential Revision: D57725201

fbshipit-source-id: 872c8104e8e35b4f429354d79f5b53c5ed7dbc0d