Remove support for prebuilt_rust_library
Summary:
`prebuilt_rust_library` is currently used for the sysroot artifacts generated as part of the tp2 rustc build. As of D54281568, those targets will no longer be necessary, so we can drop support for the prebuilt rust library.

The motivation for deleting this comes from working on `metadata-fast` in D57738355. The main issue here is that there's not a great way to actually *use* prebuilt_rust_library without guaranteeing that the dependencies marked within the prebuilt's rmeta section match the rest of the build. In other words, the prebuilt would *have* to have been built against the same stdlib as the rest of the build. At that point, you'd best be able to guarantee that by simply *not* prebuilding the artifact.

Note, for rust libs that only expose C symbols (staticlibs, cdylibs) you can still use `prebuilt_cxx_library`. If we have a need for this in the future, we can resurrect it.

Reviewed By: dtolnay

Differential Revision: D57757816

fbshipit-source-id: 4af978cbf5f2ee79d8ccd29382a24a41b1125754