Fix duplicate archive member lto prepare bug
Summary:
This diff is motivated by a bug in dist_lto_prepare.py that results in the same archive member being passed multiple times to the linker. The Darwin linker cannot handle this.

It also modifies the script to reduce the number of file rename operations required.

Here is a more detailed description of the motivating issue. Consider we have in a directory `first/foo.txt` and `second/foo.txt`, the contents of which is different. We place both in an archive as follows:

```
~/fbsource/fbcode/third-party-buck/platform010/build/llvm-fb/15/bin/llvm-ar r test.a first/foo.txt second/foo.txt
```

Then we extract them as was previously performed:

```
~/fbsource/fbcode/third-party-buck/platform010/build/llvm-fb/15/bin/llvm-ar xv test.a
x - foo.txt
x - foo.txt
```

First we notice that the output files are not unique, as we had relied upon. This resulted in the same file name being added to the manifest multiple times, which is not ok for Darwin. This diff stops relying upon the output of `llvm-ar xv` for a unique list of members.

Next, rather than renaming the foo.txt to tmp, and back to foo.txt for every duplicate member, we simply extract into a temporary directory and move the final renamed version into the objects path. This eliminates two unnecessary renames per duplicate.

One might notice that the initial `llvm-ar xv` already extracted the last duplicate and so we don't necessarily need to extract it again. That is `second/foo.txt` will be left in the cwd after the extraction above. We might be able to save an additional extraction per duplicate this way, but the fact that `llvm-ar xv` leaves behind the last duplicate, and not the first or a random copy seems like an implementation detail we don't want to rely on unless necessary. For that reason, we avoid it in this diff.

Reviewed By: rmaz

Differential Revision: D57889230

fbshipit-source-id: 2e5122baa2f100ea2bd912fbf0d938f28762e43e