buck2 | apple_resource_dedupe_alias
Summary:
# Context
To build valid XCFrameworks that expose Swift-implemented features, we need to produce a .swiftmodule directory populated with artifacts from both architectures of a split_transition universal binary

To implement this, we need to expose both sides of the split transition to the `apple_bundle_resources` implementation (D57496716). This potentially changes the build graph, resulting in `apple_bundle` rules receiving up to 3 artifacts (universal, x86_64, and arm64) for a given rule.

For most rules, this can be handled by using / extending `apple_resource_transition()`, which is already implemented and is intended to reduce the '3 artifacts' situation back down to a diamond-depdendency graph, solving the problem.

There are some non `apple_` rules that can also trigger this though.
The one case I've seen is in bundling of React Native JS artifacts via `rn_js_bundle()`.

To handle arbitrary non-apple rules like `rn_js_bundle` this diff adds `apple_resource_dedupe_alias()`, which is a special implementation of `alias()` that relies on `apple_resource_transition()` to de-dupe the resulting artifacts.

Reviewed By: blackm00n

Differential Revision: D58013618

fbshipit-source-id: 4cee48ec294aebd5d19e72cc4008da1b13bc0cac