Switch check builds to metadata-fast
Summary:
`metadata-fast` emit mode produces a lighter-weight `.rmeta` file that, importantly, excludes things like monomorphizations that are not required for type checking. This results in a build that is about 12.5% faster, when dtolnay first investigated this in D55113078.

When backing out D57625262, it's clear to see the impact:

```
$ buck2 build fbcode//smart/glassworks_meta/prod_service:prod_service_lib[check]
$ buck2 log what-ran | grep prod_service_lib
# Run the action that shows from previous command
$ time <action>
```

## Before (check [pic])
```
real    1m17.872s
user    1m14.374s
sys     0m3.521s
```

## After (check [fast])
```
real    0m0.347s
user    0m0.236s
sys     0m0.090s
```

However, the downside is that we may see less cache hit rate, so I'm not sure how to weigh that tradeoff. I'll start a post in buck2 dev on this.

Reviewed By: dtolnay

Differential Revision: D57738355

fbshipit-source-id: 963dbd4252bed692ff2549cd3a680cf34fea5a1f