Make sure com/facebook gets a license header
Summary: The two files this touches are actually shipped to open source, so need appropriate headers. While this isn't exactly the same pattern our ShipIt config uses (that looks for `facebook` not after `com`), there are many divergences between the two (keeping the two in sync would be a pain), and this gives the equivalent answer for now, so go with it.

Reviewed By: zertosh

Differential Revision: D58338526

fbshipit-source-id: 32307120982391241fe1163389b4e5e5e616d262