Add extra_xcode_files to XcodeDataInfo provider
Summary:
Several rule types (mostly apple_* and cxx_* rules) have an `extra_xcode_files` attribute, which are designed to contain files that are not part of the build we want displayed in Xcode.

Currently, to get these attributes, Xcode project generation queries the target and then extracts them from the attributes. However, these files can be generated (ie, targets) and simply querying to get the file path is not enough.

In order to clean up how we handle generated targets in the next diff, make these files available in the XcodeDataInfo provider so that project generation can easily obtain this information.

Differential Revision: D58328755

fbshipit-source-id: a8b8d8c16ba9721d1ecc26059cbce205baa2acad