Fix autodeps in fbcode for gcno provider
Summary:
`cpp_gcno_file_list` depends on definitions from prelude, which chokes autodeps.
We're going to run into this more often as we merge more fbcode_macros with UDRs and prelude.
For now, punt on this by moving the gcno provider into its own defs.bzl so we avoid the
problematic dependency.

Differential Revision: D58382992

fbshipit-source-id: 0901d33a65eacbd9a34ee6fe5536ec891ca8e206