add propagated_target_sdk_version to apple_binary
Summary: As part of migration to toolchain target_sdk_version flags, we need a way to propagate the target_sdk_version from an apple_binary. This can be done with a transition, but we cannot transition on the existing target_sdk_version attribute as transitions do not support selects on incoming attributes, which makes it impossible to introduce as is.

Reviewed By: milend

Differential Revision: D58204433

fbshipit-source-id: 0c28d950cc10a65db741a551242c6235accb61c2