Fix crate_root bug when multiple targets have the same generated crate root
Summary:
I believe this logic is not correct:

https://www.internalfb.com/code/fbsource/[bd614173b2a3928f00e01100e00947f00393769a]/fbcode/buck2/prelude/rust/rust-analyzer/resolve_deps.bxl?lines=168-172%2C175-176%2C179-181%2C183-190%2C192-193

`materialize_generated_thrift` is called once per Thrift target, and is supposed to return each of the mapped src artifacts for that target so that the caller can properly pick out the crate root.

https://www.internalfb.com/code/fbsource/[bd614173b2a3928f00e01100e00947f00393769a]/fbcode/buck2/prelude/rust/rust-analyzer/resolve_deps.bxl?lines=129%2C131-133

Multiple Thrift targets can have overlapping generated sources. This can happen in the case of plugins. The code in resolve_deps.bxl is using a set `seen` to avoid multiple `actions.copy_file` writing to the same destination, which is not allowed.

```lang=text,counterexample
  * fbcode/buck2/prelude/rust/rust-analyzer/resolve_deps.bxl:267, in expand_and_resolve_impl
      resolved_deps = cquery_deps(ctx, expanded_targets, workspaces, actions)
  * fbcode/buck2/prelude/rust/rust-analyzer/resolve_deps.bxl:129, in cquery_deps
      thrift_files = materialize_generated_thrift(ctx, target, actions, seen)
  * fbcode/buck2/prelude/rust/rust-analyzer/resolve_deps.bxl:184, in materialize_generated_thrift
      copied = actions.copy_file(outfile, artifacts.artifacts()[0])
error: Multiple artifacts and/or metadata files are declared at the same output location `fbcode/configerator/structs/data_access_policies/upf/pxl/compact_ast-rust/consts.rs` declared at `fbcode/buck2/prelude/rust/rust-analyzer/resolve_deps.bxl:184:22-74`.
   --> fbcode/buck2/prelude/rust/rust-analyzer/resolve_deps.bxl:184:22
    |
184 |             copied = actions.copy_file(outfile, artifacts.artifacts()[0])
    |                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```

However, the bug here is that when no `copy_file` action is performed, no artifact ends up in `out` and the caller will be unable to set a correct crate root. If a project depends on one of those affected targets for which bxl traversal order reaches a different target earlier which uses the same generated crate root, then rust-analyzer will not be able to build it.

Reviewed By: davidbarsky

Differential Revision: D58631999

fbshipit-source-id: 564f9dc67d51049a5d4a30b4e114800686fa27e5