Support macro interpolation in named_deps names
Summary:
`named_deps` is a map of crate name to Rust library target.

```
rust_library(
    ...
    named_deps = {
        "clap5": "//third-party/rust:clap-5.0.0-alpha.0",
    },
)
```

Previously, string macros like `$(location ...)` were not supported in the keys of this map. This diff adds support for it, which I'll need in D58537300 for dynamically selecting non-conflicting crate names for all the Thrift dependencies of a generated Thrift library target.

```
rust_library(
    ...
    named_deps = {
        "$(crate_name //path/to:cratename)_services": "//path/to:rust-library",
    },
)
```

This is a much less invasive change than {D44933138} because we never need to use the keys of `named_deps` in any filenames. All that needs to happen is for them to get shuffled through to an `--extern {cratename}=path/to/rlib` rustc argument, which Buck `attrs.arg()` is well suited for.

Reviewed By: diliop

Differential Revision: D58537292

fbshipit-source-id: 12f560d092e231a315c1a6d19c4da9c45df6f00e