Support resolved macros in named deps
Summary:
The change in {D58537292} makes the `named_deps` attribute of `rust_binary` and `rust_library` able to accept string interpolations, like this:

```
rust_library(
    ...
    named_deps = {
        "$(crate_name //path/to:cratename)_services": "//path/to:rust-library",
    },
)
```

That other diff contains everything that's needed to support this feature in builds performed by Buck, but for builds performed by rust-analyzer, that diff only has a small hack in resolve_deps.bxl to make rust-analyzer continue to function like before as long as no `named_deps` containing interpolations are actually used in fbcode.

This diff fills in the parts of the implementation required for rust-analyzer to work with interpolations in `named_deps` crate names.

The approach is that in the JSON output of resolved_deps.bxl, instead of producing named deps as a dictionary mapping analysis-time crate name to target label as before:

```
-    "named_deps": {
-      "thrift_protocol": "fbcode//thrift/lib/thrift:protocol_detail-rust",
-      "thrift_schema": "fbcode//thrift/lib/thrift:schema-rust"
-    },
```

we'll instead produce a build artifact containing the true build-time value for N crate names, followed by the N target labels for those crates in the same order:

```
+    "named_deps": [
+      "/data/users/dtolnay/fbsource/buck-out/v2/gen/fbcode/99cab575c71b0a4b/ligen/compiler/thrift/__adapter__/named_deps",
+      "fbcode//thrift/lib/thrift:protocol_detail-rust",
+      "fbcode//thrift/lib/thrift:schema-rust"
+    ],
```

```
$ cat /data/users/dtolnay/fbsource/buck-out/v2/gen/fbcode/99cab575c71b0a4b/ligen/compiler/thrift/__adapter__/named_deps
thrift_protocol
thrift_schema
```

Then in `rust-project`'s JSON deserialization code, we can read that named_deps file and build the same FxHashMap that would have been deserialized from a JSON map. This is nice because the rest of rust-project outside of deserialization needs no changes.

Reviewed By: zertosh

Differential Revision: D58638036

fbshipit-source-id: 4005a1028a59952836062440792183d080012176