Add `coverage_instrumentation_compiler_flags` attribute
Summary:
Add `coverage_instrumentation_compiler_flags` attribute to `cxx_library` and
`cxx_binary` rules. This attribute will be set always and will contain the
compilation flags that are required to instrument sources. Prelude code will only used them when coverage is enabled for a target due to a header being selected for coverage.

Reviewed By: Nekitosss

Differential Revision: D57494358

fbshipit-source-id: 88c3d7f93505250b943f2cdc16b74bb8344fbf74