Only Adding Actually Used Hidden Directories
Summary: We're making every src directory for the dependencies as hidden, meaning we're stating that we need them symlinked for the script to work. However, we don't look at the source directories unless the module is included in the list of deps. So, this only marks actually used directories as hidden.

Reviewed By: blackm00n

Differential Revision: D58610247

fbshipit-source-id: 4ee6be7576b1d34eb4fb3156ec758712638dc9d6