Setting Action Weight
Summary:
This action is failing on continuous builds occasionally; seemingly randomly.

The only clue as to what's going wrong is there's a status code 139, which could indicate a SIGSEGV.

Trying to set a weight for the action here thinking that just maybe the problem is we're running out of memory and saying that running less of these actions in parrellel will avoid this.

Reviewed By: chatura-atapattu

Differential Revision: D58792819

fbshipit-source-id: 7d2524ccbfe59674d7cfb823d73075f43e8e3ef7