prebuilt_apple_framework | sanitizer for distribution
Summary:
When using a framework or xcframework in a target, we need to use the headers and modules files to compile.

However, when distributing the framework as part of an app bundle, these files are just bloat.

For convenience, in this diff we offer a 'distribution' subtarget of apple_prebuilt_framework() that contains the framework with those files removed. It supports versioned (i.e. macOS) and un versioned (i.e. iOS) frameworks.

Reviewed By: milend

Differential Revision: D58815797

fbshipit-source-id: 10885282040e6f949c7c5a19ed9a5835d38d5683