Delete support for --extern-html-root-url
Summary:
No longer used as of {D59138761}.

Use `rustc_flags += ["-Zcrate-attr=doc(html_root_url = {})".format(...)]` in a macro instead.

If we reinstate an `html_root_url_prefix` attribute on the RustToolchainInfo in the future, we might want to consider doing it in a way that supports different prefixes per cell, or by package prefix.

```
html_root_url_prefix = {
    "fbcode": "/intern/rustdoc",
}
```

Reviewed By: zertosh

Differential Revision: D59138763

fbshipit-source-id: fa1b516ec4f2612007f2bae29060df5c82562306