Fix Swift wrapper
Summary:
The fix for the Swift compiler wrapper to handle missing file outputs was only outputting in case stdout/stderr contained `could not write`. This regressed the behaviour of the wrapper, so that:

- Missing stdout/stderr from the underlying invocation if the Swift compiler fails for any other reasons (e.g., syntax error, compiler error etc)
- stdout/stderr now merged into a single stream (stdout)

Fix the shell script, so that we *always* output and don't merge into a single output stream.

Reviewed By: chatura-atapattu

Differential Revision: D59268935

fbshipit-source-id: 12261733986514705ce0aa093b6ab920caaf00d1