Allow overriding the `python_binary` output extension in the target
Summary:
Right now, the extension is configured by the `python_toolchain`. However, this is a global override whereas we may need a per-target override for the extension.

This allows a `python_binary` to delcare an `extension` attribute to set its own output.

According to the Buck1 code, `python_binary.extension` overrides the `pexExtension` default.

Reviewed By: JakobDegen

Differential Revision: D59244979

fbshipit-source-id: 1f63a6ea94b6a6b3b2ad9d7d0da7d7be6664d38b