erlang: migrate app_scr_builder to use json instead of terms
Summary:
This migrates the app builder script to using json instead of erlang terms for communication.
This should be more efficient.

On the Erlang side, this also fixes indeterminism in the job - there were maps with atom keys used converted to list with `maps:to_list` or printed with just `~p` - this can cause non-determinism. Instead this diff tries to make those values sorted.

Reviewed By: TheGeorge

Differential Revision: D59285894

fbshipit-source-id: c9d15842b20cd50594078eaaed03f3aa7f4b0964