erlang: migrate test runner to json from term specs
Summary:
This switches test runner from using term test specs to using JSON. This should help lower memory pressure from buck2 when running tests since constructing the test specs as terms turned out pretty expensive in other cases.

The other changes are related to the fact that now we pass-in data as binaries rather than strings.
In particular the `erl` command was concatenated into a single string on buck side, only to be re-split on the erlang side - instead we pass it in as an array of strings now.

Additionally, I switched code loading of hooks to be more assertive - before it would fail silently and then result in suspicious failures - failing early makes it much easier to understand what's going on.

Writing the test spec file was simplified to do just one IO operation, rather than a sequence of writes for key config value.

Reviewed By: TD5

Differential Revision: D59328376

fbshipit-source-id: 9c9e3373babea50741c89bbddf3315104847f0ed