Use common opt argsfile
Summary:
We use a response file to store arguments for "opt actions" in distributed
thin-lto builds. Prior to this diff, we created a new response file for every
opt action, of which there may be thousands. All the reponse files are identical
since all opt actions for a given link unit use the same flags. We can just write
the response file once and re-use it.

Reviewed By: rmaz

Differential Revision: D59840768

fbshipit-source-id: 19a2c6b2ee2fa118277ab4bc2edab65028be0ee1